A myriad of changes, still working on gettting the other builds to pass however. Will be factoring out embedded compiler references so the interpreter can run on it's own
