+++ src/core/latria_char.h
@@ -1,0 +1,43 @@
+/*
+ The MIT License (MIT)
+ 
+ Copyright (c) 2017 Benjamin Wilson Friedman
+ 
+ Permission is hereby granted, free of charge, to any person obtaining a copy
+ of this software and associated documentation files (the "Software"), to deal
+ in the Software without restriction, including without limitation the rights
+ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ copies of the Software, and to permit persons to whom the Software is
+ furnished to do so, subject to the following conditions:
+ 
+ The above copyright notice and this permission notice shall be included in all
+ copies or substantial portions of the Software.
+ 
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ SOFTWARE.
+ */
+
+/*
+ *  latria_char.h
+ */
+
+#ifndef latria_core_latria_char_h
+#define latria_core_latria_char_h
+
+#include "latria_bool.h"
+
+/* Indicates whether this character is a recognized operator */
+LATBool isCharacterOperator(char c);
+
+/* Expression Checking Expression */
+char isArithmeticOperator(char i);
+
+/* Indicates whether this is a recognized whitespace character */
+char isWhitespace(char c);
+
+#endif /* latria_core_latria_char_h */
