+++ src/memory/latria_memory_block.h
@@ -1,0 +1,51 @@
+/*
+ The MIT License (MIT)
+ 
+ Copyright (c) 2017 Benjamin Wilson Friedman
+ 
+ Permission is hereby granted, free of charge, to any person obtaining a copy
+ of this software and associated documentation files (the "Software"), to deal
+ in the Software without restriction, including without limitation the rights
+ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ copies of the Software, and to permit persons to whom the Software is
+ furnished to do so, subject to the following conditions:
+ 
+ The above copyright notice and this permission notice shall be included in all
+ copies or substantial portions of the Software.
+ 
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ SOFTWARE.
+ */
+
+/*
+ *  latria_memory_block.h
+ */
+
+#ifndef latria_memory_latria_memory_block_h
+#define latria_memory_latria_memory_block_h
+
+#include "core/latria_core.h"
+
+/* Latria's Block Chain */
+struct MemoryBlock {
+    
+    /* The next block in the chain */
+    struct MemoryBlock *nextBlock;
+    
+    /* The start of the memory block this points to */
+    void *blockPointer;
+    
+    /* The size of this block */
+    size_t blockSize;
+    
+    /* Marker for whether this block has been allocated or not */
+    unsigned char isAllocated;
+    
+};
+
+#endif /* latria_memory_latria_memory_block_h */
